backend_django/requirements.web.txt
Django==5.2.9
djangorestframework
django-extensions
python-dotenv
psycopg[binary]

celery
redis

minio>=7.2.0
qdrant-client

Pillow

sentence-transformers
openai
transformers
PyMuPDF
drf-spectacular
drf-spectacular-sidecar